GUI Commit
While I do have the UI working, merging the styled form and the audio interactivity is insanely difficult merely due to the fact that Blazor is a web app.
But I have learnt how to create counters which will allow me to communicate between the form and the web app. Which means it is possible